lab 2/Lab_work_2/Image_Vectorization/requirements.txt
# pip install -r requirements.txt
# pip freeze
# pip uninstall -r requirements.txt
# pip freeze >> requirements.txt

contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
scipy==1.15.3
six==1.16.0
